REF: change ExtractEnumVariant from an intention to a refactoring
